Move the storybook helper
